aergoluac: separate luac module from util
